#  [pdata] [chore] Use extracted copyOrig instead of CopyTo on new objects