--- a/pdata/pcommon/generated_float64slice.go
+++ b/pdata/pcommon/generated_float64slice.go
@@ -40,12 +40,12 @@ func NewFloat64Slice() Float64Slice {
 
 // AsRaw returns a copy of the []float64 slice.
 func (ms Float64Slice) AsRaw() []float64 {
-	return copyFloat64Slice(nil, *ms.getOrig())
+	return internal.CopyOrigFloat64Slice(nil, *ms.getOrig())
 }
 
 // FromRaw copies raw []float64 into the slice Float64Slice.
 func (ms Float64Slice) FromRaw(val []float64) {
-	*ms.getOrig() = copyFloat64Slice(*ms.getOrig(), val)
+	*ms.getOrig() = internal.CopyOrigFloat64Slice(*ms.getOrig(), val)
 }
 
 // Len returns length of the []float64 slice value.
@@ -98,10 +98,5 @@ func (ms Float64Slice) MoveTo(dest Float64Slice) {
 
 // CopyTo copies all elements from the current slice overriding the destination.
 func (ms Float64Slice) CopyTo(dest Float64Slice) {
-	*dest.getOrig() = copyFloat64Slice(*dest.getOrig(), *ms.getOrig())
-}
-
-func copyFloat64Slice(dst, src []float64) []float64 {
-	dst = dst[:0]
-	return append(dst, src...)
+	*dest.getOrig() = internal.CopyOrigFloat64Slice(*dest.getOrig(), *ms.getOrig())
 }