--- a/pdata/pcommon/slice.go
+++ b/pdata/pcommon/slice.go
@@ -66,17 +66,7 @@ func (es Slice) At(ix int) Value {
 
 // CopyTo copies all elements from the current slice overriding the destination.
 func (es Slice) CopyTo(dest Slice) {
-	srcLen := es.Len()
-	destCap := cap(*dest.getOrig())
-	if srcLen <= destCap {
-		(*dest.getOrig()) = (*dest.getOrig())[:srcLen:destCap]
-	} else {
-		(*dest.getOrig()) = make([]otlpcommon.AnyValue, srcLen)
-	}
-
-	for i := range *es.getOrig() {
-		newValue(&(*es.getOrig())[i]).CopyTo(newValue(&(*dest.getOrig())[i]))
-	}
+	internal.CopyOrigSlice(dest.getOrig(), es.getOrig())
 }
 
 // EnsureCapacity is an operation that ensures the slice has at least the specified capacity.