This change slightly improves performance of CopyTo methods and makes things easier for https://github.com/open-telemetry/opentelemetry-collector/pull/7205

Before:
```
goos: darwin
goarch: arm64
pkg: go.opentelemetry.io/collector/pdata/pmetric
BenchmarkMetricsCopyTo
BenchmarkMetricsCopyTo-10    	     890	   1299537 ns/op	 2286313 B/op	   42359 allocs/op
BenchmarkLogsCopyTo
BenchmarkLogsCopyTo-10    	   50179	     23463 ns/op	   66400 B/op	     513 allocs/op
BenchmarkTracesCopyTo
BenchmarkTracesCopyTo-10    	    3712	    352776 ns/op	  613639 B/op	    6687 allocs/op
BenchmarkMetricsCopyToReuseDest
BenchmarkMetricsCopyToReuseDest-10    	     898	   1248538 ns/op	 2252874 B/op	   42189 allocs/op
BenchmarkLogsCopyToReuseDest
BenchmarkLogsCopyToReuseDest-10    	  234370	      5152 ns/op	       0 B/op	       0 allocs/op
BenchmarkTracesCopyToReuseDest
BenchmarkTracesCopyToReuseDest-10    	   23239	     50635 ns/op	      26 B/op	       0 allocs/op
PASS
```

After:
```
goos: darwin
goarch: arm64
pkg: go.opentelemetry.io/collector/pdata/pmetric
BenchmarkMetricsCopyTo
BenchmarkMetricsCopyTo-10    	     928	   1229692 ns/op	 2286312 B/op	   42359 allocs/op
BenchmarkLogsCopyTo
BenchmarkLogsCopyTo-10    	   51979	     23356 ns/op	   66400 B/op	     513 allocs/op
BenchmarkTracesCopyTo
BenchmarkTracesCopyTo-10    	    3144	    345044 ns/op	  613637 B/op	    6687 allocs/op
BenchmarkMetricsCopyToReuseDest
BenchmarkMetricsCopyToReuseDest-10    	     986	   1189443 ns/op	 2252871 B/op	   42189 allocs/op
BenchmarkLogsCopyToReuseDest
BenchmarkLogsCopyToReuseDest-10    	  249128	      4534 ns/op	       0 B/op	       0 allocs/op
BenchmarkTracesCopyToReuseDest
BenchmarkTracesCopyToReuseDest-10    	   26362	     44933 ns/op	      23 B/op	       0 allocs/op
PASS
```