@dmitryax 2023-02-19
Memory allocations in the Metrics.CopyTo benchmark seem concerning. I'll take a look into that